Move Folders and Add Eraser Feature
Moved folders around and created an eraser feature for the level editor. When the eraser tool is being used, you can click prefabs in the scene to delete it.
