--- Assets/Scripts/Optimization.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 524ced93edb97ce4fbbe2c35c925c043
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
