+++ Assets/Scripts/Tools/MergeMeshesManager.cs
