+++ Assets/Scripts/Tools/MergeMeshesManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 4a596be1e4486c04ebbf5aeba2b24581
+guid: dd656e43d15d2ea42a24f414a25ff8ad
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
